Factor out into iam.ts
